requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow==2.10.0
python-multipart
